+++ build-env/assets/static/js/exposed.js
@@ -1,3 +1,3 @@
 function emit(event, data) {
-  return axios.post(`/emit/${event}`, data);
+  return axios.post(`/emit/${event}`, data ?? {});
 }