+++ build.ps1
@@ -18,8 +18,3 @@ npx ncc build .\dist\AsphyxiaCore.js -o .\build-env --external pug --external ts
 # Packing
 Write-Output "Packing exe"
 npx pkg .\build-env -t "node12.16.1-win-x64,node12.16.1-win-x86" -o .\build\asphyxia-core --options no-warnings
-
-# Packing zips
-Compress-Archive -Path ".\build\asphyxia-core-win-x64.exe", ".\plugins" -DestinationPath ".\build\asphyxia-core-win-x64.zip" -Force
-Compress-Archive -Path ".\build\asphyxia-core-win-x86.exe", ".\plugins" -DestinationPath ".\build\asphyxia-core-win-x86.zip" -Force
-Compress-Archive -Path ".\build\asphyxia-core-linux-x64", ".\plugins" -DestinationPath ".\build\asphyxia-core-linux-x64.zip" -Force